DASH-1344 Separate tests into own files.
