+++ src/tests/constants.ts
@@ -1,0 +1,2 @@
+export const MEMO_PHRASE = 'Hello FIO SDK Lite';
+export const MEMO_PHRASE_ENCRYPTED = 'Hello FIO SDK Lite Encrypted';
