+++ src/tests/types.d.ts
@@ -1,0 +1,5 @@
+export type TestWallet = {
+  privateKey: string;
+  publicKey: string;
+  fioHandle: string;
+};
